+++ Tests/SentryTests/Protocol/SentryMetaTests.swift
@@ -1,0 +1,27 @@
+import XCTest
+
+class SentryMetaTests: XCTestCase {
+
+    override func tearDown() {
+        SentryMeta.clearSdkPackages()
+        SentrySdkPackage.setSentryPackageInfoForTests(3) //Unknown (default)
+    }
+
+    func testPackagesAreNotNil() {
+        let packages = SentryMeta.getSdkPackages()
+        XCTAssertNotNil(packages)
+        XCTAssertEqual(0, packages.count)
+    }
+
+    func testPackagesIncludeSdkPackage() {
+        SentrySdkPackage.setSentryPackageInfoForTests(0) //SPM
+
+        let packages = SentryMeta.getSdkPackages()
+        XCTAssertEqual(1, packages.count)
+
+        let package = packages.first!
+        XCTAssertEqual("spm:getsentry/sentry.cocoa", package.name)
+        XCTAssertEqual(SentryMeta.versionString, package.version)
+    }
+
+}